# specs/001-llm-server-test-harness/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Test Harness API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - ApiTokenAuth: []
components:
  securitySchemes:
    ApiTokenAuth:
      type: apiKey
      in: header
      name: X-API-Token
  schemas:
    Target:
      type: object
      required: [id, name, base_url]
      properties:
        id: { type: string }
        name: { type: string }
        base_url: { type: string }
        auth_type: { type: string }
        default_model: { type: string }
        default_params: { type: object }
        timeouts: { type: object }
        concurrency_limit: { type: integer }
...
    Run:
      type: object
      required: [id, status, started_at]
      properties:
        id: { type: string }
        target_id: { type: string }
        test_id: { type: string }
        suite_id: { type: string }
        profile_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed, canceled] }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
    TestResult:
      type: object
      required: [id, run_id, test_id, verdict]
      properties:
        id: { type: string }
        run_id: { type: string }
        test_id: { type: string }
        verdict: { type: string, enum: [pass, fail, skip] }
        failure_reason: { type: string }
        metrics: { type: object }
        artefacts: { type: object }
        raw_events: { type: array, items: { type: object } }
...
paths:
...
  /targets:
    get:
      summary: List targets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Target' }
    post:
      summary: Create target
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Target' }
      responses:
        '201': { description: Created }
  /targets/{targetId}:
    get:
      summary: Get target
      parameters:
        - in: path
          name: targetId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      summary: Update target
      parameters:
        - in: path
          name: targetId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Target' }
      responses:
        '200': { description: OK }
    delete:
      summary: Delete target
      parameters:
        - in: path
          name: targetId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
  /runs:
    post:
      summary: Start a run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_id]
              properties:
                target_id: { type: string }
                test_id: { type: string }
                suite_id: { type: string }
                profile_id: { type: string }
                profile_version: { type: string }
                repetitions: { type: integer }
                continue_on_failure: { type: boolean }
      responses:
        '202': { description: Accepted }
    get:
      summary: List runs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Run' }
  /runs/{runId}:
    get:
      summary: Get run
      parameters:
        - in: path
          name: runId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /runs/{runId}/results:
    get:
      summary: List results for a run
      parameters:
        - in: path
          name: runId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TestResult' }
...